Refactor ReferralStepCard with inner presentational component
